+++ modules/kube-modules/outputs.tf
@@ -1,6 +1,0 @@
-output "example" {
-  description = "Example output"
-  value       = <<-EOH
-    This is just a test-message!
-  EOH
-}