+++ modules/kube-modules/variables.tf
@@ -2,7 +2,7 @@
 #   REQUIRED (passed from parent module)
 # ==============================================================================
 variable "enabled"     { description = "See parent module" }
-variable "kube_config" { description = "See parent module" }
+#variable "kube_config" { description = "Credentials for accessing the deployed aks cluster" }
 # ---
 variable "admin_group"        { description = "See parent module" }
 variable "dashboard_as_admin" { description = "See parent module" }