+++ test/main.tf
@@ -16,18 +16,22 @@ resource "random_string" "prefix" {
 # Deployment
 
 module "testing_advanced" {
-  #source       = "github.com/www-aiqu-no/terraform-azurerm-aks?ref=master"
-  #version      = "0.1.0"
-
-  #See https://www.terraform.io/docs/configuration/modules.html#passing-providers-explicitly
-  #providers = {}
-
   source       = "../."
   prefix       = random_string.prefix.result
   cluster_type = "advanced"
   #initialized  = true
 
-  #log_analytics_enabled   = true
-  #kube_management_enabled = true
+  log_analytics_enabled   = true
+  kube_management_enabled = true
+  kube_dashboard_as_admin = true
   #kube_admin_group = "<Group Id or Name>"
+}
+
+# NOTE: To print output from modules, you need to export it in root:
+output "kube_info" {
+  value = module.testing_advanced.info
+}
+
+output "kube_public_ssh_key" {
+  value = module.testing_advanced.public_ssh_key
 }